requirements.txt
crewai[anthropic]>=0.28.0
crewai-tools>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
litellm>=1.0.0

